+++ cpp/subprojects/boosting/src/mlrl/boosting/rule_evaluation/rule_evaluation_decomposable_partial_dynamic_binned.cpp
@@ -30,8 +30,8 @@ namespace boosting {
         protected:
 
             uint32 calculateOutputWiseCriteria(const StatisticVector& statisticVector, float64* criteria,
-                                               uint32 numCriteria, float64 l1RegularizationWeight,
-                                               float64 l2RegularizationWeight) override {
+                                               uint32 numCriteria, float32 l1RegularizationWeight,
+                                               float32 l2RegularizationWeight) override {
                 uint32 numElements = statisticVector.getNumElements();
                 typename StatisticVector::const_iterator statisticIterator = statisticVector.cbegin();
                 const std::pair<float64, float64> pair =
@@ -79,8 +79,8 @@ namespace boosting {
             DecomposableDynamicPartialBinnedRuleEvaluation(const IndexVector& labelIndices,
                                                            std::unique_ptr<PartialIndexVector> indexVectorPtr,
                                                            float32 threshold, float32 exponent,
-                                                           float64 l1RegularizationWeight,
-                                                           float64 l2RegularizationWeight,
+                                                           float32 l1RegularizationWeight,
+                                                           float32 l2RegularizationWeight,
                                                            std::unique_ptr<ILabelBinning> binningPtr)
                 : AbstractDecomposableBinnedRuleEvaluation<StatisticVector, PartialIndexVector>(
                     *indexVectorPtr, true, l1RegularizationWeight, l2RegularizationWeight, std::move(binningPtr)),
@@ -89,7 +89,7 @@ namespace boosting {
     };
 
     DecomposableDynamicPartialBinnedRuleEvaluationFactory::DecomposableDynamicPartialBinnedRuleEvaluationFactory(
-      float32 threshold, float32 exponent, float64 l1RegularizationWeight, float64 l2RegularizationWeight,
+      float32 threshold, float32 exponent, float32 l1RegularizationWeight, float32 l2RegularizationWeight,
       std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr)
         : threshold_(threshold), exponent_(exponent), l1RegularizationWeight_(l1RegularizationWeight),
           l2RegularizationWeight_(l2RegularizationWeight), labelBinningFactoryPtr_(std::move(labelBinningFactoryPtr)) {}
